+++ js/time/time.controller.js
@@ -2,8 +2,8 @@
 //
 // TIME
 //
-angular.module("cycloneApp").controller("TimeCtrl", ["$scope", "Auth", "$firebaseArray", "focus", "$timeout", "$rootScope", "$route", "moment",
-    function($scope, Auth, $firebaseArray, focus, $timeout, $rootScope, $route, moment) {
+angular.module("cycloneApp").controller("TimeCtrl", ["$scope", "Auth", "$firebaseArray", "focus", "$timeout", "$rootScope", "$route", "moment", "timeTypesService",
+    function($scope, Auth, $firebaseArray, focus, $timeout, $rootScope, $route, moment, timeTypesService) {
         // Angular-clipboard
         $scope.copySuccess = function () {
             console.log('Copied time!');
@@ -14,7 +14,8 @@ angular.module("cycloneApp").controller("TimeCtrl", ["$scope", "Auth", "$firebas
         };
 
         // The different types
-        $scope.types = ['work', 'internal', 'private'];
+        $scope.types = timeTypesService.getTimeTypes();
+        // TODO: Make this a configuration option or save it in the firebasedb
 
         // check the route when ready
         $rootScope.$on('$routeChangeSuccess', function () {
@@ -108,7 +109,6 @@ angular.module("cycloneApp").controller("TimeCtrl", ["$scope", "Auth", "$firebas
 
             // Duration
             $scope.currentDuration = 0;
-            $scope.currentDurationHasHours = false;
             $scope.lastEntryTimestamp = lastEntryTimestamp; // So the first entry works
 
             // Show the current date
@@ -303,12 +303,13 @@ angular.module("cycloneApp").controller("TimeCtrl", ["$scope", "Auth", "$firebas
                 }
                 if ($scope.newContinueEntryType !== undefined){
                     $scope.newEntryType = $scope.newContinueEntryType;
-                    $scope.newContinueEntryType = ''; // Clear it again
+                    $scope.newContinueEntryType = 'work'; // Default it again
                 }
 
                 // Focus First element now again, so we are ready to type an other task
                 focus('newTaskProject');
 
+                // TODO: put the following logic into a function to get called also by other change options later
                 // Which id and location did we save the entry? We need to check the prev and next entry to update the duration!
                 var newEntryKey = queryRef.key;
                 // Get location in the array
@@ -451,12 +452,6 @@ angular.module("cycloneApp").controller("TimeCtrl", ["$scope", "Auth", "$firebas
         // Realtime duration display
         function updateDurations() {
             $scope.currentDuration = ((Date.now() - $scope.lastEntryTimestamp));
-            // Set the HasHours expression
-            if (((1000 * 60 * 60) - $scope.currentDuration) <= 0) {
-                $scope.currentDurationHasHours = true;
-            } else {
-                $scope.currentDurationHasHours = false;
-            }
             $timeout(updateDurations, 1000, true);
         };